fix(audit): pass 3 — keep settingscontroller under 800 lines, stale comment ref
- settingscontroller.cpp crossed the 800-line cap (779 at v3.1 → 814) when this PR
  added the snapping sub-controller ctor wiring; extract the active-page/dirty-
  tracking/external-edit methods into settingscontroller_pagestate.cpp (686 now),
  matching the established sibling-TU split pattern
- WindowTrackingService.cpp: replace another brittle cross-file line-number comment
  ("saveload.cpp:257-259") with a symbolic reference to the adaptor save/load behavior
