+++ src/settings/settingscontroller_pagestate.cpp
@@ -1,0 +1,155 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+// Active-page navigation + per-page dirty tracking + external-edit envelope
+// for SettingsController:
+//   * setActivePage   — switch the viewed page (with parent→leaf redirect)
+//   * onSettings/ExternalSettingsChanged — NOTIFY → dirty / reload hooks
+//   * setNeedsSave / dirtyPages / isPageDirty — dirty-state surface for QML
+//   * begin/endExternalEdit — stack envelope so sidebar/global widgets mark
+//     the correct page dirty
+//
+// Split out of settingscontroller.cpp to keep that file under the 800-line
+// cap. Same class, separate TU, no API change.
+
+#include "settingscontroller.h"
+
+#include "../core/logging.h"
+
+#include <QDebug>
+
+namespace PlasmaZones {
+
+void SettingsController::setActivePage(const QString& page)
+{
+    // Resolve parent category names (e.g. "snapping" → "snapping-zones")
+    const QString resolved = parentPageRedirects().value(page, page);
+
+    if (!validPageNames().contains(resolved)) {
+        qCWarning(PlasmaZones::lcCore) << "Unknown settings page:" << page;
+        return;
+    }
+    // Reentrancy guard: a slot connected to activePageChanged that
+    // calls setActivePage again (e.g. a CLI --page handler that
+    // redirects to a fallback page) would otherwise re-trigger
+    // m_loading toggling, leaving the toggle in an unspecified state
+    // if the inner call set m_loading = false before the outer call's
+    // restore ran. Returning early on re-entry keeps m_loading's
+    // false→true→false window symmetric per public-entry call.
+    if (m_settingActivePage) {
+        qCWarning(PlasmaZones::lcCore) << "setActivePage: reentrant call refused (already setting active page to"
+                                       << m_activePage << ")";
+        return;
+    }
+    if (m_activePage != resolved) {
+        // m_loading suppresses onSettingsPropertyChanged — the QML Loader
+        // reacts synchronously to activePageChanged and new page creation
+        // may trigger NOTIFY signals that would otherwise mark pages dirty.
+        m_settingActivePage = true;
+        m_loading = true;
+        m_activePage = resolved;
+        Q_EMIT activePageChanged();
+        m_loading = false;
+        m_settingActivePage = false;
+    }
+}
+
+void SettingsController::onSettingsPropertyChanged()
+{
+    if (!m_saving && !m_loading) {
+        setNeedsSave(true);
+    }
+}
+
+void SettingsController::onExternalSettingsChanged()
+{
+    if (!m_saving) {
+        load();
+    }
+}
+
+void SettingsController::setNeedsSave(bool needs)
+{
+    // Mark the target page as dirty, or clear all dirty pages if needs ==
+    // false. The target is the top of the external-edit stack when set
+    // (sidebar / global widgets that mutate settings owned by a different
+    // page than the one the user is viewing), otherwise m_activePage.
+    // Parent categories ("snapping", "tiling") are never the active page —
+    // setActivePage redirects them to their first child — so the target
+    // always resolves to a concrete leaf page.
+    if (needs) {
+        const QString target = m_externalEditStack.isEmpty() ? m_activePage : m_externalEditStack.top();
+        Q_ASSERT(!parentPageRedirects().contains(target));
+        if (!m_dirtyPages.contains(target)) {
+            m_dirtyPages.insert(target);
+            Q_EMIT dirtyPagesChanged();
+        }
+    } else if (!m_dirtyPages.isEmpty()) {
+        m_dirtyPages.clear();
+        Q_EMIT dirtyPagesChanged();
+    }
+}
+
+QStringList SettingsController::dirtyPages() const
+{
+    // Order is unspecified — QML uses this only as a binding dependency
+    // and calls isPageDirty() for the actual lookup.
+    return QStringList(m_dirtyPages.begin(), m_dirtyPages.end());
+}
+
+bool SettingsController::isPageDirty(const QString& page) const
+{
+    if (m_dirtyPages.contains(page))
+        return true;
+    // Parent / virtual-parent category: dirty if any child leaf in
+    // the group is dirty. Single direct-membership lookup against
+    // `pageGroupChildren()` rather than the old prefix-walk-or-hash-
+    // lookup branch — top-level parents (snapping / tiling /
+    // animations) and virtual mid-level parents (animations-surfaces /
+    // animations-library) share the same code path now.
+    const auto& groups = pageGroupChildren();
+    const auto it = groups.constFind(page);
+    if (it != groups.constEnd()) {
+        for (const QString& child : *it) {
+            if (m_dirtyPages.contains(child))
+                return true;
+        }
+    }
+    return false;
+}
+
+void SettingsController::beginExternalEdit(const QString& page)
+{
+    // Resolve parent categories to their canonical leaf — same rules as
+    // setActivePage — so the sidebar can pass "snapping" or "tiling".
+    const QString resolved = parentPageRedirects().value(page, page);
+    if (!validPageNames().contains(resolved)) {
+        qCWarning(PlasmaZones::lcCore) << "beginExternalEdit: unknown page" << page;
+        return;
+    }
+    // Push onto the stack so nested begin/end pairs restore the outer
+    // target on pop instead of clearing the wrap entirely. This is
+    // genuinely reachable: an animations-page pendingChangesChanged
+    // handler can fire synchronously while the controller is inside a
+    // window-rules-driven external-edit envelope, and the inner pair
+    // must not erase the outer target.
+    m_externalEditStack.push(resolved);
+}
+
+void SettingsController::endExternalEdit()
+{
+    if (m_externalEditStack.isEmpty()) {
+        // Defence-in-depth: an unmatched end means a begin was lost or
+        // a caller is double-popping. Warn so the failure is visible
+        // instead of silently no-oping (the previous QString-clear
+        // form was equally silent, but a stack pop on empty would
+        // crash in debug builds without this guard).
+        qCWarning(PlasmaZones::lcCore) << "endExternalEdit: stack is empty — unmatched end?";
+        Q_ASSERT_X(false, "SettingsController::endExternalEdit",
+                   "endExternalEdit called with no matching beginExternalEdit on the stack.");
+        return;
+    }
+    m_externalEditStack.pop();
+}
+
+} // namespace PlasmaZones
